test/mochitest/test_install_risky.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="progressDialog"><div>
    <section>
      <h2>Flashing Device</h2>
      <p>Do not unplug your device while flashing…</p>
      <ul>
        <li class="downloading">Downloading Build</li>
        <li class="extracting">Extracting Build</li>
        <li class="fetching">Downloading blobs from device</li>
        <li class="creating">Creating final build</li>
        <li class="flashing">Flashing build</li>
      </ul>
      <div id="additionalProgress"></div>
      <div id="subAdditionalProgress"></div>
    </section>
  </div></div>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  var realDownloadProgress, realDownloadBuild, realStep,
      realDistributionStep, realDeviceStep, realImageStep, realFlashStep;

  function test_setup_fakeenv() {
    return new Promise((resolve, reject) => {
      is(typeof downloadProgress, "function", "downloadProgress function");
      is(typeof downloadBuild, "function", "downloadBuild function");
      is(typeof distributionStep, "function", "distributionStep function");
      is(typeof deviceStep, "function", "deviceStep function");
      is(typeof imageStep, "function", "imageStep function");
      is(typeof flashStep, "function", "flashStep function");
      realDownloadProgress = downloadProgress;
      realDownloadBuild    = downloadBuild;
      realDistributionStep = distributionStep;
      realDeviceStep       = deviceStep;
      realImageStep        = imageStep;
      realFlashStep        = flashStep;

      resolve();
    });
  }

  function test_execute_resolve() {
    return new Promise((resolve, reject) => {
      console.debug("resolving ...");
      assert_isrisky();
      resolve();
    });
  }

  function test_execute_reject() {
    return new Promise((resolve, reject) => {
      console.debug("rejecting ...");
      assert_isrisky();
      reject();
    });
  }

  function test_unsetup_fakeenv() {
    return new Promise((resolve, reject) => {
      downloadProgress = realDownloadProgress;
      downloadBuild    = realDownloadBuild;
      distributionStep = realDistributionStep;
      deviceStep       = realDeviceStep;
      imageStep        = realImageStep;
      flashStep        = realFlashStep;
      resolve();
    });
  }

  function assert_not_isrisky() {
    is(_isRisky, false, "is not in risky state");
  }

  function assert_isrisky() {
    is(_isRisky, true, "is in risky state");
  }

  function runTest() {
    loadAboutJs()
      .then(() => test_setup_fakeenv())
      .then(() => assert_not_isrisky())
      .then(() => {
        downloadProgress = test_execute_resolve;
        downloadBuild    = test_execute_resolve;
        distributionStep = test_execute_resolve;
        deviceStep       = test_execute_resolve;
        imageStep        = test_execute_resolve;
        flashStep        = test_execute_resolve;
      })
      .then(() => install())
      .then(() => assert_not_isrisky())
      .then(() => {
        downloadBuild    = test_execute_reject;
      })
      .then(() => install())
      .catch(() => assert_not_isrisky())
      .then(() => test_unsetup_fakeenv())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
